+++ yash-builtin/src/cd.rs
@@ -21,7 +21,7 @@
 //! # Synopsis
 //!
 //! ```sh
-//! cd [-L|-P] [directory]
+//! cd [-L|-P [-e]] [directory]
 //! ```
 //!
 //! # Description
@@ -76,6 +76,12 @@
 //! These two options are mutually exclusive. The last specified one applies if
 //! given both. The default is `-L`.
 //!
+//! When the `-P` option is effective, the built-in may fail to determine the
+//! new working directory pathname to assign to `$PWD`. By default, the exit
+//! status does not indicate the failure. If the **`-e`** (**`--ensure-pwd`**)
+//! option is given together with the `-P` option, the built-in returns exit
+//! status 1 in this case.
+//!
 //! TODO: The **`--default-directory=directory`** option is not implemented.
 //!
 //! TODO: The **`--print={always,auto,never}`** option is not implemented.
@@ -88,12 +94,24 @@
 //!
 //! # Errors
 //!
-//! This built-in may fail with a non-zero exit status when:
+//! This built-in fails if the working directory cannot be changed, for example,
+//! in the following cases:
 //!
 //! - The operand does not resolve to an existing accessible directory.
 //! - The operand is omitted and `$HOME` is not set or empty.
 //! - The operand is a single hyphen (`-`) and `$OLDPWD` is not set or empty.
 //! - The resolved pathname of the new working directory is too long.
+//!
+//! If the `-P` option is effective, the built-in may fail to determine the
+//! new working directory pathname to assign to `$PWD`, for example, in the
+//! following cases:
+//!
+//! - The new pathname is too long.
+//! - Some ancestor directories of the new working directory are not accessible.
+//! - The new working directory does not belong to the filesystem tree.
+//!
+//! In these cases, the working directory remains changed, the `$PWD` variable
+//! is left empty, and the exit status depends on the `-e` option.
 //!
 //! The built-in may also fail in the following cases, but the working directory
 //! will remain changed and the exit status will be zero:
@@ -103,7 +121,19 @@
 //!
 //! # Exit Status
 //!
-//! Zero if the working directory was successfully changed; non-zero otherwise.
+//! - If the working directory is successfully changed:
+//!   - If the `-L` option is effective, the exit status is zero.
+//!   - If the `-L` option is effective:
+//!     - If the new working directory pathname is successfully determined, the
+//!       exit status is zero.
+//!     - If the new working directory pathname cannot be determined:
+//!       - If the `-e` option is effective, the exit status is one.
+//!       - Otherwise, the exit status is zero.
+//! - If the working directory cannot be changed because of an error in the
+//!   underlying `chdir` system call, the exit status is two.
+//! - If the operand cannot be processed because of an unset or empty `$HOME` or
+//!   `$OLDPWD`, the exit status is three.
+//! - If the command arguments are invalid, the exit status is four.
 //!
 //! # Security considerations
 //!
@@ -114,11 +144,11 @@
 //! the working directory to an unexpected one. To ensure that the cd built-in
 //! behaves as intended, shell script writers should unset the variable at the
 //! beginning of the script. Users can configure `$CDPATH` in their shell
-//! sessions, but must avoid exporting the variable to the environment.
+//! sessions, but should avoid exporting the variable to the environment.
 //!
 //! # Portability
 //!
-//! The `-L` and `-P` options are defined in POSIX. The other options are
+//! The `-L`, `-P`, and `-e` options are defined in POSIX. The other options are
 //! non-standard.
 //!
 //! The shell sets `$PWD` on the startup and modifies it in the cd built-in.
@@ -128,6 +158,9 @@
 //! The error handling behavior and the exit status do not agree between
 //! existing implementations when the built-in fails because of a write error or
 //! a read-only variable error.
+//!
+//! Other implementations may return different non-zero exit statuses in cases
+//! where this implementation would return exit statuses between 2 and 4.
 //!
 //! POSIX requires the shell to convert the pathname passed to the underlying
 //! `chdir` system call to a shorter relative pathname when the `-L` option is
@@ -144,13 +177,33 @@
 //! the above conditions are not met. The current implementation does it if and
 //! only if the final operand starts with `$PWD`.
 
-use crate::common::report_error;
-use crate::common::report_failure;
+use crate::common::report;
 use crate::Result;
 use yash_env::path::Path;
+use yash_env::path::PathBuf;
+use yash_env::semantics::ExitStatus;
 use yash_env::semantics::Field;
+use yash_env::system::Errno;
 use yash_env::variable::PWD;
 use yash_env::Env;
+use yash_syntax::source::pretty::AnnotationType;
+use yash_syntax::source::pretty::Footer;
+use yash_syntax::source::pretty::Message;
+
+/// Exit status when the built-in succeeds
+pub const EXIT_STATUS_SUCCESS: ExitStatus = ExitStatus(0);
+
+/// Exit status when the new `$PWD` value cannot be determined
+pub const EXIT_STATUS_STALE_PWD: ExitStatus = ExitStatus(1);
+
+/// Exit status for an error in the underlying `chdir` system call
+pub const EXIT_STATUS_CHDIR_ERROR: ExitStatus = ExitStatus(2);
+
+/// Exit status for an unset or empty `$HOME` or `$OLDPWD`
+pub const EXIT_STATUS_UNSET_VARIABLE: ExitStatus = ExitStatus(3);
+
+/// Exit status for invalid command arguments
+pub const EXIT_STATUS_SYNTAX_ERROR: ExitStatus = ExitStatus(4);
 
 /// Treatments of symbolic links in the pathname
 #[derive(Debug, Clone, Copy, Default, Eq, Hash, PartialEq)]
@@ -169,7 +222,15 @@ pub enum Mode {
 #[non_exhaustive]
 pub struct Command {
     /// Treatments of symbolic links in the pathname
+    ///
+    /// The `-L` and `-P` options are translated to this field.
     pub mode: Mode,
+
+    /// Whether to ensure the new `$PWD` value
+    ///
+    /// The `-e` option is translated to this field.
+    /// This option must be used together with the `-P` option.
+    pub ensure_pwd: bool,
 
     /// The operand that specifies the directory to change to
     pub operand: Option<Field>,
@@ -188,20 +249,42 @@ fn get_pwd(env: &Env) -> String {
     env.variables.get_scalar(PWD).unwrap_or_default().to_owned()
 }
 
+/// Reports that the new `$PWD` value cannot be determined, and returns the
+/// corresponding exit status.
+async fn report_pwd_error(env: &mut Env, errno: Errno, ensure_pwd: bool) -> Result {
+    let (r#type, exit_status) = if ensure_pwd {
+        (AnnotationType::Error, EXIT_STATUS_STALE_PWD)
+    } else {
+        (AnnotationType::Warning, EXIT_STATUS_SUCCESS)
+    };
+
+    let message = Message {
+        r#type,
+        title: "cannot compute new $PWD".into(),
+        annotations: vec![],
+        footers: vec![Footer {
+            r#type: AnnotationType::Info,
+            label: format!("error from underlying system call: {errno}").into(),
+        }],
+    };
+
+    report(env, message, exit_status).await
+}
+
 /// Entry point for executing the `cd` built-in
 ///
 /// This function uses functions in the submodules to execute the built-in.
 pub async fn main(env: &mut Env, args: Vec<Field>) -> Result {
     let command = match syntax::parse(env, args) {
         Ok(command) => command,
-        Err(e) => return report_error(env, &e).await,
+        Err(e) => return report(env, &e, EXIT_STATUS_SYNTAX_ERROR).await,
     };
 
     let pwd = get_pwd(env);
 
     let (path, origin) = match target::target(env, &command, &pwd) {
         Ok(target) => target,
-        Err(e) => return report_failure(env, &e).await,
+        Err(e) => return report(env, &e, EXIT_STATUS_UNSET_VARIABLE).await,
     };
 
     let short_path = shorten::shorten(&path, Path::new(&pwd), command.mode);
@@ -211,11 +294,59 @@ pub async fn main(env: &mut Env, args: Vec<Field>) -> Result {
         Err(e) => return chdir::report_failure(env, command.operand.as_ref(), &path, &e).await,
     }
 
-    let new_pwd = assign::new_pwd(env, command.mode, &path);
+    let (new_pwd, result) = match assign::new_pwd(env, command.mode, &path) {
+        Ok(new_pwd) => (new_pwd, Result::from(EXIT_STATUS_SUCCESS)),
+        Err(errno) => (
+            PathBuf::default(),
+            report_pwd_error(env, errno, command.ensure_pwd).await,
+        ),
+    };
+
     print::print_path(env, &new_pwd, &origin).await;
 
     assign::set_oldpwd(env, pwd).await;
     assign::set_pwd(env, new_pwd).await;
 
-    Result::default()
+    result
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use futures_util::FutureExt as _;
+    use std::rc::Rc;
+    use yash_env::VirtualSystem;
+    use yash_env_test_helper::assert_stderr;
+
+    #[test]
+    fn report_pwd_error_with_ensure_pwd() {
+        let system = Box::new(VirtualSystem::new());
+        let state = Rc::clone(&system.state);
+        let mut env = Env::with_system(system);
+
+        let result = report_pwd_error(&mut env, Errno::ENAMETOOLONG, true)
+            .now_or_never()
+            .unwrap();
+
+        // Something should be printed
+        assert_stderr(&state, |stderr| assert!(!stderr.is_empty()));
+
+        assert_eq!(result, Result::from(ExitStatus(1)));
+    }
+
+    #[test]
+    fn report_pwd_error_without_ensure_pwd() {
+        let system = Box::new(VirtualSystem::new());
+        let state = Rc::clone(&system.state);
+        let mut env = Env::with_system(system);
+
+        let result = report_pwd_error(&mut env, Errno::ENAMETOOLONG, false)
+            .now_or_never()
+            .unwrap();
+
+        // Something should be printed
+        assert_stderr(&state, |stderr| assert!(!stderr.is_empty()));
+
+        assert_eq!(result, Result::from(ExitStatus(0)));
+    }
 }